ADD: Implement records management for patients; add routes and controller methods for retrieving and adding records; remove old records controller
