+++ api/src/controllers/recordings.ts
@@ -20,9 +20,11 @@ const addRecording = async (
     const filePath = `recordings/${body.patientId}/${fileName}.${body.fileType}`;
     await s3.addFile(filePath, body.data);
 
-    await axios.post(`${process.env.AI_API_URL}process`, {
+    const request = await axios.post(`${process.env.AI_API_URL}process`, {
       filePath,
     });
+
+    console.log(request.data);
 
     res.json({ success: true });
   } catch (error) {
